requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
asttokens==2.2.1
async-timeout==4.0.2
attrs==23.1.0
backcall==0.2.0
backoff==2.2.1
black==23.3.0
blinker==1.6.2
boto3==1.26.134
botocore==1.29.134
cachetools==5.3.0
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
comm==0.1.3
contourpy==1.0.7
cycler==0.11.0
debugpy==1.6.7
decorator==5.1.1
dnspython==2.3.0
executing==1.2.0
Flask==2.3.2
Flask-Cors==3.0.10
fonttools==4.39.4
frozenlist==1.3.3
google-api-core==2.11.0
google-auth==2.18.0
google-cloud-core==2.3.2
google-cloud-storage==2.9.0
google-cloud-vision==3.4.1
google-crc32c==1.5.0
google-resumable-media==2.5.0
googleapis-common-protos==1.59.0
gql==3.4.1
graphql-core==3.2.3
grpcio==1.54.2
grpcio-status==1.54.2
idna==3.4
importlib-metadata==6.6.0
importlib-resources==5.12.0
ipykernel==6.23.1
ipython==8.13.2
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
jupyter-client==8.2.0
jupyter-core==5.3.0
kiwisolver==1.4.4
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
multidict==6.0.4
mypy==1.3.0
mypy-extensions==1.0.0
nest-asyncio==1.5.6
numpy==1.24.3
packaging==23.1
parso==0.8.3
pathspec==0.11.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.5.1
prompt-toolkit==3.0.38
proto-plus==1.22.2
protobuf==4.23.0
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
Pygments==2.15.1
PyJWT==2.7.0
pymongo==4.3.3
pynamodb==5.5.0
pyparsing==3.0.9
python-dateutil==2.8.2
pyzmq==25.0.2
requests==2.30.0
requests-toolbelt==1.0.0
rsa==4.9
s3transfer==0.6.1
scikit-learn==1.2.2
scipy==1.10.1
six==1.16.0
stack-data==0.6.2
threadpoolctl==3.1.0
tomli==2.0.1
tornado==6.3.2
traitlets==5.9.0
types-protobuf==4.23.0.1
typing-extensions==4.5.0
urllib3==1.26.15
wcwidth==0.2.6
Werkzeug==2.3.4
yarl==1.9.2
zipp==3.15.0
